+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.1.0 (June 15, 2025)
 
+* Added `Delete Object By ID` Action
 * Added `Lookup Object By ID` Action
 
 # 1.0.0 (June 03, 2025)